Remove VPC endpoint for API Gateway
This HTTP 403 error most commonly occurs when private DNS is enabled for an API Gateway interface VPC endpoint that's associated with an Amazon VPC. When this happens, all requests from the VPC to API Gateway APIs resolve to that interface VPC endpoint. However, it's not possible to connect to public APIs using a VPC endpoint.